Fix test\jest\Save.test.ts
